Killaura

+++ Echo's Client/Math/Math.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <Windows.h>
+#include <cmath>
 
 class Math {
 public:
@@ -41,11 +42,11 @@ struct Vec3 {
 	bool operator == (Vec3 v) { return v.x == x && v.y == y && v.z == z; };
 	bool operator != (Vec3 v) { return v.x != x || v.y != y || v.z != z; };
 
-	double distance(Vec3 v) {
+	float distance(Vec3 v) {
 		float dX = x - v.x;
 		float dY = y - v.y;
 		float dZ = z - v.z;
-		return sqrt(dX * dX + dY * dY + dZ * dZ);
+		return std::sqrt(dX * dX + dY * dY + dZ * dZ);
 	}
 };
